+++ src/failprint/_internal/types.py
@@ -1,9 +1,4 @@
-"""Special types.
-
-Attributes:
-    CmdType: Type for a command.
-    CmdFuncType: Type for a command or function.
-"""
+# Special types.
 
 from __future__ import annotations
 
@@ -12,6 +7,8 @@ from typing import Callable, Union
 from failprint._internal.lazy import LazyCallable
 
 CmdType = Union[str, list[str]]
+"""Type for a command."""
 CmdFuncType = Union[CmdType, Callable, LazyCallable]
+"""Type for a command or function."""
 
 __all__ = ["CmdFuncType", "CmdType"]